Clean up - Refactor resource controller - Part 1
