+++ incubator/virtualcluster/pkg/syncer/resources/namespace/controller.go
@@ -17,31 +17,23 @@ limitations under the License.
 package namespace
 
 import (
-	"fmt"
-	"strings"
-
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	coreinformers "k8s.io/client-go/informers/core/v1"
 	v1core "k8s.io/client-go/kubernetes/typed/core/v1"
 	listersv1 "k8s.io/client-go/listers/core/v1"
-	"k8s.io/client-go/tools/cache"
 	"k8s.io/klog"
 
-	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/constants"
-	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/conversion"
 	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/manager"
 	mc "github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/mccontroller"
-	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/reconciler"
 )
 
 type controller struct {
-	namespaceClient                 v1core.NamespacesGetter
-	multiClusterNamespaceController *mc.MultiClusterController
-
+	// super master namespace client
+	namespaceClient v1core.NamespacesGetter
+	// super master namespace lister
 	nsLister listersv1.NamespaceLister
-	nsSynced cache.InformerSynced
+	// Connect to all tenant master namespace informers
+	multiClusterNamespaceController *mc.MultiClusterController
 }
 
 func Register(
@@ -61,90 +53,16 @@ func Register(
 		return
 	}
 	c.multiClusterNamespaceController = multiClusterNamespaceController
-
 	c.nsLister = namespaceInformer.Lister()
-	c.nsSynced = namespaceInformer.Informer().HasSynced
 
 	controllerManager.AddController(c)
 }
 
 func (c *controller) StartUWS(stopCh <-chan struct{}) error {
-	if !cache.WaitForCacheSync(stopCh, c.nsSynced) {
-		return fmt.Errorf("failed to wait for caches to sync")
-	}
-
 	return nil
-}
-
-func (c *controller) StartDWS(stopCh <-chan struct{}) error {
-	return c.multiClusterNamespaceController.Start(stopCh)
 }
 
 func (c *controller) StartPeriodChecker(stopCh <-chan struct{}) {
-}
-
-func (c *controller) Reconcile(request reconciler.Request) (reconciler.Result, error) {
-	klog.Infof("reconcile namespace %s %s event for cluster %s", request.Name, request.Event, request.Cluster.Name)
-
-	switch request.Event {
-	case reconciler.AddEvent:
-		err := c.reconcileNamespaceCreate(request.Cluster.Name, request.Name, request.Obj.(*v1.Namespace))
-		if err != nil {
-			klog.Errorf("failed reconcile namespace %s CREATE of cluster %s %v", request.Name, request.Cluster.Name, err)
-			return reconciler.Result{Requeue: true}, nil
-		}
-	case reconciler.UpdateEvent:
-		err := c.reconcileNamespaceUpdate(request.Cluster.Name, request.Name, request.Obj.(*v1.Namespace))
-		if err != nil {
-			klog.Errorf("failed reconcile namespace %s UPDATE of cluster %s %v", request.Name, request.Cluster.Name, err)
-			return reconciler.Result{}, err
-		}
-	case reconciler.DeleteEvent:
-		err := c.reconcileNamespaceRemove(request.Cluster.Name, request.Name)
-		if err != nil {
-			klog.Errorf("failed reconcile namespace %s DELETE of cluster %s %v", request.Name, request.Cluster.Name, err)
-			return reconciler.Result{}, err
-		}
-	}
-	return reconciler.Result{}, nil
-}
-
-func (c *controller) reconcileNamespaceCreate(cluster, name string, namespace *v1.Namespace) error {
-	targetNamespace := conversion.ToSuperMasterNamespace(cluster, name)
-	_, err := c.nsLister.Get(targetNamespace)
-	if err == nil {
-		// namespace is ready.
-		return nil
-	}
-
-	newObj, err := conversion.BuildSuperMasterNamespace(cluster, namespace)
-	if err != nil {
-		return err
-	}
-
-	_, err = c.namespaceClient.Namespaces().Create(newObj.(*v1.Namespace))
-	if errors.IsAlreadyExists(err) {
-		klog.Infof("namespace %s of cluster %s already exist in super master", name, cluster)
-		return nil
-	}
-	return err
-}
-
-func (c *controller) reconcileNamespaceUpdate(cluster, name string, namespace *v1.Namespace) error {
-	return nil
-}
-
-func (c *controller) reconcileNamespaceRemove(cluster, name string) error {
-	targetName := strings.Join([]string{cluster, name}, "-")
-	opts := &metav1.DeleteOptions{
-		PropagationPolicy: &constants.DefaultDeletionPolicy,
-	}
-	err := c.namespaceClient.Namespaces().Delete(targetName, opts)
-	if errors.IsNotFound(err) {
-		klog.Warningf("namespace %s of cluster %s not found in super master", name, cluster)
-		return nil
-	}
-	return err
 }
 
 func (c *controller) AddCluster(cluster mc.ClusterInterface) {
